I've unlocked the LED code completely, all of the LED's can be individually toggled.
In addition, the two LED's for "ON" (Blue1) and "START" (Green2) will now blink at 1Hz.

So, yes they can be toggled, but they will change again within the next second if they are indicating the respective states.

The opcode error LED, Red3, will blink until cleared with one of its LED commands.

There is now a separate LED for Sub-Command errors, Red2, which will behave the same as the Opcode Error LED.


Signed-off-by: PaladinoJ <PaladinoJ@gmail.com>