Made some significant progress on odometry processing.
Calibrated the readOdom() function in proteus.py to return the correct steering angle.  This was determined by finding the ACTUAL turning radius and figuring out what the steering angle for it would be.

Based on Steering Angle and the distances between wheels, have written functions that can translate a move into cartesian coordinates.

Signed-off-by: Justin Paladino <PaladinoJ@gmail.com>